refactor : 메인페이지 물병 컴포넌트 리팩토링 및 애니메이션 추가
